patch 0.0.2
Patch 0.0.2: The Effects patch

This patch brings joy, because it improves the effects system, making it graphics based, not instance based. This should increase fps drastically when dealing with a lot of effects.

Saving and loading saves has been temporarily disabled due to maintenance, but I should deal with it next patch.

And that theres still no options menu.